added dynamic giphy sdk flow
