+++ likeminds-chat-reactnative-integration/ChatSX/optionalDependecies/Gif.ts
@@ -1,0 +1,9 @@
+let GIFPicker: any;
+
+try {
+  GIFPicker = require("@giphy/react-native-sdk");
+} catch (err) {
+  console.log("@giphy/react-native-sdk is not installed");
+}
+
+export default GIFPicker;
